pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "magi-system"
version = "0.1.0"
description = "MAGI multi-agent decision system scaffold with RAG, DSPy personas, and evaluation workflows."
readme = "magi/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "MAGI Team"}]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",
  "pypdf>=5.0",
  "numpy>=1.26",
  "scikit-learn>=1.3",
  "httpx>=0.27",
  "psycopg[binary]>=3.2"
]

[project.optional-dependencies]
openai = ["openai>=1.40.0"]
google = ["google-genai>=0.4.0"]
dspy = ["dspy>=2.5.0"]
torch = ["torch>=2.2"]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5.0",
  "mypy>=1.10",
  "types-PyYAML>=6.0"
]

[project.scripts]
magi = "magi.app.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["magi", "magi.*"]

[tool.pytest.ini_options]
testpaths = ["magi/tests"]
addopts = ["--strict-config", "--strict-markers"]

[tool.ruff]
target-version = "py310"
